fix bug in hungarian algorithm alignment
if we want to map i -> j, the permutation matrix needs a 1 at (j, i), whereas
previously I was putting it at (i, j). Also, when working with a wide matrix,
the permutation matrix will also be a subsetting matrix, so there is no longer a
need to pad the diagonals.

I updated the test to cover this new case and confirm that it is doing
assignment correctly