fix critical bug in alignment
CCA is, in some senses, scale invariant: the canonical correlations are
preserved under rescaling of the input variables (it's even more invariant than
that; I think they are preserved under arbitrary full rank transformations of
the vectors). However, the directions are *not* invariant, which matters for
things like alignment.

I was trying to have my cake and eat it, too. I kept everything on the original
scale of the data, so that constructing canonical variates would work nicely,
but then I stored away the original standard deviations in the cancor object
when doing cancor_scaled. The idea was that when doing alignment (and
specifically when calculating the cosine similarity of two solutions), I'd tweak
the directions so that they corresponded to what one would get if they had done
per-variable standardization prior to running CCA.

I naively implemented this by *dividing* the canonical directions by the
standard deviations, when I should have been *multiplying* them. As variables
increase in scale, their standard deviation goes up. As a result, the
corresponding canonical directions get smaller (to offset this and preserve the
correlation invariance). Thus, standardizing a big variable would reduce the
standard deviation but increase the direction, so I should be multiplying rather
than dividing!

This likely had pretty wacky and severe consequences, since it was then
de-emphasizing variables with high variance, which often correspond to where I
put the signal in my simulation studies. This has implications for essentially
all types of alignment that operate on cosine similarity, which is almost all of
them.

This also makes me realize that the posmax alignment should also take this into
account, since we don't want the underlying scale of the variables to influence
where the max shows up.