rework signflip test
signflip should be based on averaged cosine similarity, not the result of
concatenating the vectors (as it presently does, but this will get fixed in a
subsequent commit). the current behavior would allow the larger vector to
dominate the cosine similarity, whereas taking the average guards against this.
however, due to the standardization inherent in cosine similarity, the current
test case would not actually yield a flip, so put thumb on scale to ensure that
it does