thoroughly rework all alignment strategies
this is a big change that includes refactoring as well as change in intended
functionality.

first, abstract standardization and trimming into separate functions, then
incorporate it as a preprocessing step in all alignment functions as
appropriate.

then, rework all the functions to actually use the trimmed and standardized
values whenever making comparisons. also ensure that anything that permutes
directions also permutes correlations, if applicable.

rework all strategies based on cosine similarity to use the average rather than
the concatenated vector.

all current tests pass, but coverage is not great so this requires some smoke testing