add de.bias.CCA as a dependency (my dev version)
Add the package from Laha et al as a dependency so we can use it in simulation
studies.

VERY IMPORTANT: the package repo is called de.bias.CCA, but the package NAME (as
defined in the description) is de.bias.cca; note the difference in case.

The installation process is case sensitive. It seems like devtools which
actually query (or perhaps even download?) all of the remotes listed and look at
their DESCRIPTION files to see what packages they provide, so it's important to
get the case write in the Imports: field